+++ kartothek/core/index.py
@@ -845,9 +845,13 @@ def _index_dct_to_table(index_dct: IndexDictType, column: str, dtype: pa.DataTyp
     # Additional note: pyarrow.array is supposed to infer type automatically.
     # But the inferred type is not enough to hold np.uint64. Until this is fixed in
     # upstream Arrow, we have to retain the following line
-    keys = np.array(list(keys))
+    if not index_dct:
+        # the np.array dtype will be double which arrow cannot convert to the target type, so use an empty list instead
+        labeled_array = pa.array([], type=dtype)
+    else:
+        keys = np.array(list(keys))
+        labeled_array = pa.array(keys, type=dtype)
 
-    labeled_array = pa.array(keys, type=dtype)
     partition_array = pa.array(list(index_dct.values()))
 
     return pa.Table.from_arrays(